% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creates_PAD_profile.R
\name{creates_PAD_profile}
\alias{creates_PAD_profile}
\title{Conversion to 1D profile}
\usage{
creates_PAD_profile(
  matrix,
  dtm = NULL,
  vox_size = 0.3,
  plot_area = 1600,
  normalize = F,
  debug = F,
  export = F,
  output_dir = NULL,
  plot_name = NULL
)
}
\arguments{
\item{matrix}{A 4 columns dataframe containing coordonates of each point with its VegDens value (X, Y, Z, VegDens).}

\item{dtm}{raster. Digital Terrain Model for the study area. Necessary if normalize = TRUE.}

\item{vox_size}{numeric. Dimension of the voxels/grid (in meter). default = 0.3 m}

\item{plot_area}{numeric. Dimension of the matrix, the study area (in square meter). default = 1600 m2}

\item{normalize}{logical. Choose TRUE if you want to normalize the matrix and you have a raster dtm of the study area. Uses the height_above_ground() function of lidR package.}

\item{debug}{logical. For development only.}

\item{export}{logical. Choose TRUE and specify a directory (output_dir) to export the 1D profile.}

\item{output_dir}{character. Output directory to export the intermediate results of the function.}

\item{plot_name}{character. Name of your data, for proper export procedure.}
}
\value{
Returns a 2 columns dataframe (1D profile) describing the VegDens vertical distribution (z, VegDens).
}
\description{
Converts a 3D matrix into a 1D vertical profile.
3D matrix is a 4 columns dataframe describing the VegDens distribution (X, Y, Z, VegDens).
}
\author{
Amélie Juckler
}
\keyword{internal}
