% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_DVV.R
\name{export_DVV}
\alias{export_DVV}
\title{Export results}
\usage{
export_DVV(ForestLayersList, output_dir = NULL, plot_name = NULL)
}
\arguments{
\item{ForestLayersList}{R list. Contains the return of VerticalLayers_from_1Dprofile() or VerticalLayers_from_3Dgrid() function.}

\item{output_dir}{character. Output directory to export the final results of the main function.}

\item{plot_name}{character. Name of your data, for proper export procedure.}
}
\value{
Exports the results of layers delineation and parameterization into an Excel file + the visual representation of the vertical layers of your study area (png).
}
\description{
Main function to export the final results of the package.
}
\seealso{
See VerticalLayers_from_1Dprofile() and VerticalLayers_from_3Dgrid() for utilization examples.
}
\author{
Amélie Juckler
}
\keyword{export}
