% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_limits.R
\name{final_limits}
\alias{final_limits}
\title{Get final boundaries of the vertical layers}
\usage{
final_limits(
  weibullList,
  col_name = "VegDens",
  print = F,
  debug = F,
  export = F,
  output_dir = NULL,
  plot_name = NULL
)
}
\arguments{
\item{weibullList}{R list. Contains the return of weibull_fit() function.}

\item{col_name}{character. Name of the value of vegetation density (e.g. PAD, biomass, voxel count). Used for visualization purpose. Default is "VegDens".}

\item{print}{logical. Choose TRUE to visualize intermediate graphs created during the function.}

\item{debug}{logical. For development only.}

\item{export}{logical. Choose TRUE and specify a directory (output_dir) to export the 1D profile.}

\item{output_dir}{character. Output directory to export the intermediate results of the function.}

\item{plot_name}{character. Name of your data, for proper export procedure.}
}
\value{
Returns a list of 25 columns data.frames with the VegDens estimations, structure parameters and final layers limits.
}
\description{
Redefines each vegetation layer to match the Weibull estimations.
weibullList is a list of 25 columns data.frame.
}
\author{
Amélie Juckler
}
\keyword{internal}
