% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_final_results.R
\name{get_final_results}
\alias{get_final_results}
\title{Formating results}
\usage{
get_final_results(finalList)
}
\arguments{
\item{finalList}{R list. Contains the return of final_limits() function.}
}
\value{
Returns, for each layer, a 3 columns data.frame with the height, the initial VegDens and the estimated VegDens, and a list of structure parameters.
}
\description{
Function that gives the results of the main function in their final format.
Dividing the data.frame into a data.frame with the initial VegDens and estimated VegDens, and a list of structure parameters.
}
\author{
Amélie Juckler
}
\keyword{internal}
