% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_weibull.R
\name{my_weibull}
\alias{my_weibull}
\title{Weibull function}
\usage{
my_weibull(params, x)
}
\arguments{
\item{params}{List of Weibull parameters (4).}

\item{x}{Height values.}
}
\value{
Returns the VegDens estimations with Weibull parameters for each x value.
}
\description{
Weibull function
}
\author{
Amélie Juckler
}
\keyword{internal}
