% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrmse_global.R
\name{nrmse_global}
\alias{nrmse_global}
\title{Statistics computation}
\usage{
nrmse_global(finalList)
}
\arguments{
\item{finalList}{R list. Contains the return of final_limits() function.}
}
\value{
Final NRMSE and R2 of the Weibull estimations for the full vertical profile.
}
\description{
NRMSE and R2 calculation for the full vertical profile.
}
\author{
Amélie Juckler
}
\keyword{internal}
