% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments_PAD_profile.R
\name{segments_PAD_profile}
\alias{segments_PAD_profile}
\title{Automatic identification of vertical layers in the 1D profile}
\usage{
segments_PAD_profile(
  profile,
  diff_threshold = 0.25,
  diff_max_threshold = 0.25,
  col_name = "VegDens",
  print = F,
  debug = F,
  export = F,
  output_dir = NULL,
  plot_name = NULL
)
}
\arguments{
\item{profile}{A 2 columns dataframe describing the VegDens vertical distribution (Z, VegDens). Can be the result of the creates_PAD_profile() function.}

\item{diff_threshold}{numeric. The difference threshold (between 0 and 1) corresponding to a percentage of the segment’s maximum VegDens. Default is 0.25. See future paper for more details.}

\item{diff_max_threshold}{numeric. The difference threshold (between 0 and 1) corresponding to a percentage of the profile’s maximum VegDens. Default is 0.25. See future paper for more details.}

\item{col_name}{character. Name of the value of vegetation density (e.g. PAD, biomass, voxel count). Used for visualization purpose. Default is "VegDens".}

\item{print}{logical. Choose TRUE to visualize intermediate graphs created during the function.}

\item{debug}{logical. For development only.}

\item{export}{logical. Choose TRUE and specify a directory (output_dir) to export the 1D profile.}

\item{output_dir}{character. Output directory to export the intermediate results of the function.}

\item{plot_name}{character. Name of your data, for proper export procedure.}
}
\value{
Returns a list of subtable of profile. Each subtable is a 2 columns dataframe describing the VegDens vertical distribution for one layer (z, VegDens).
}
\description{
Subdivides a VegDens vertical profile into different vegetation layers.
}
\author{
Amélie Juckler
}
\keyword{internal}
