% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VerticalLayers_from_1Dprofile.R
\name{VerticalLayers_from_1Dprofile}
\alias{VerticalLayers_from_1Dprofile}
\title{Main ForestLayers function, use with 1D data}
\usage{
VerticalLayers_from_1Dprofile(
  profile,
  diff_threshold = 0.25,
  diff_max_threshold = 0.25,
  col_name = "VegDens",
  print = FALSE,
  debug = FALSE,
  export = FALSE,
  output_dir = NULL,
  plot_name = NULL
)
}
\arguments{
\item{profile}{A 2 columns dataframe describing the VegDens vertical distribution (Z, VegDens). Can be the intermediate result of the VerticalLayers_from_3Dgrid() function.}

\item{diff_threshold}{numeric. The difference threshold (between 0 and 1) corresponding to a percentage of the segment’s maximum VegDens. Default is 0.25. See future paper for more details.}

\item{diff_max_threshold}{numeric. The difference threshold (between 0 and 1) corresponding to a percentage of the profile’s maximum VegDens. Default is 0.25. See future paper for more details.}

\item{col_name}{character. Name of the value of vegetation density (e.g. PAD, biomass, voxel count). Used for visualization purpose. Default is "VegDens".}

\item{print}{logical. Choose TRUE to visualize intermediate graphs created during the function.}

\item{debug}{logical. For development only.}

\item{export}{logical. Choose TRUE and specify a directory (output_dir) to export the 1D profile.}

\item{output_dir}{character. Output directory to export the intermediate results of the function.}

\item{plot_name}{character. Name of your data, for proper export procedure.}
}
\value{
Returns, for each layer, a 3 columns data.frame with the height, the initial VegDens and the estimated VegDens, and a list of structural parameters. Returns also a visual representation of the full parameterized vertical profile.
}
\description{
Automatic layers delineation and parameterization from a 1D vertical profile.
Converts a 1D vertical profile (2 columns data.frame) into a list of the parameterized layers with their parameters and Weibull estimations.
The results are, for each layer, a data.frame with the initial VegDens and estimated VegDens, and a list of structural parameters.
}
\examples{
\dontrun{
# Open your 2 columns data.frame and check that you have the correct columns (height, VegDens)
profile <- readxl::read_xlsx(my_file, sheet = "Sheet1")
profile <- as.data.frame(profile)

# Define ForestLayers parameters
diff_threshold = 0.25
diff_max_threshold = 0.25
col_name <- "VegDens (units)"
output_dir <- "C:/my/output/directory"
plot_name <- "my_study_area"

my_layers <- VerticalLayers_from_1Dprofile(
profile,
diff_threshold,
diff_max_threshold,
col_name,
print = FALSE,
debug = FALSE,
export = TRUE,
output_dir,
plot_name
)

# Export results
results <- export_DVV(ForestLayersList = my_layers, output_dir, plot_name)
}

}
\seealso{
Use VerticalLayers_from_3Dgrid() function if you have 3D data (grid, voxelized point cloud).
See export_DVV() function to export your results.
}
\author{
Amélie Juckler and other contributors
}
\keyword{Weibull}
\keyword{layer,}
\keyword{profile,}
\keyword{vertical}
