% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_function.R
\name{visu}
\alias{visu}
\title{Visualization of the full vertical profile}
\usage{
visu(finalList, col_name = "VegDens", debug = F)
}
\arguments{
\item{finalList}{R list. Contains the return of final_limits() function.}

\item{col_name}{character. Name of the value of vegetation density (e.g. PAD, biomass, voxel count). Used for visualization purpose. Default is "VegDens".}

\item{debug}{logical. For development only.}
}
\value{
Returns a plot with the raw vertical profile of the study area and the fitted Weibull distributions for each layer.
}
\description{
Creates the visual representation of the vertical layers of your study area.
}
\author{
Amélie Juckler
}
\keyword{internal}
