% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_fit.R
\name{weibull_fit}
\alias{weibull_fit}
\title{Fits a Weibull distribution on each vertical layer}
\usage{
weibull_fit(segmentList, col_name = "VegDens", print = FALSE, debug = FALSE)
}
\arguments{
\item{segmentList}{R list. Contains the return of segments_PAD_profile() function.}

\item{col_name}{character. Name of the value of vegetation density (e.g. PAD, biomass, voxel count). Used for visualization purpose. Default is "VegDens".}

\item{print}{logical. Choose TRUE to visualize intermediate graphs created during the function.}

\item{debug}{logical. For development only.}
}
\value{
Returns a list of subtable of profile. Each subtable is a 25 columns data.frame describing the VegDens vertical distribution, the Weibull distribution and the structural parameters for one layer.
}
\description{
Fits Weibull distribution on the VegDens profile in each vegetation layer.
segmentList is a list of 2 columns data.frame corresponding to the VegDens profile of different vegetation layers (x = Z, y = VegDens).
}
\author{
Amélie Juckler
}
\keyword{internal}
